+++ repositories/anydatabase.go
@@ -1,7 +1,7 @@
 package repositories
 
 import (
-	"Dp218Go/services"
+	"Dp218Go/services/usecases"
 	"context"
 	"github.com/jackc/pgconn"
 	"github.com/jackc/pgx/v4"
@@ -13,5 +13,5 @@ type AnyDatabase interface {
 	QueryExec(context.Context, string, ...interface{}) (pgconn.CommandTag, error)
 	CloseDB()
 
-	services.UserUsecases
+	usecases.UserUsecases
 }
